// gestion-patrimoine-front-end/src/app/enum/etat-bon-pour.enum.ts
export enum EtatBonPour {

  INITIAL = 'INITIAL',
  BAF = 'EN COURS D\'ENREGISTREMENT',
  ALLERDLF = 'TRANSMIS AU DLF',
  ALLERBLM = 'TRANSMIS AU BLM',
  ALLERSECTION = 'TRANSMIS A LA SECTION',
  RETOURSECTION = 'EN COURS DE TRAITEMENT',
  RETOURBLM = 'RETRANSMIS AU BLM',
  RETOURDLF = 'RETRANSMIS AU DLF',
  TERMINER = 'TERMINER'
  
}